Rename .js to .mjs
