+++ lib/js/components/layouts/ramp/NodeSpecPropertiesManager.typeroof.jsx
@@ -1,11 +1,11 @@
 import { _BaseByPathContainerComponent } from "./_BaseByPathContainerComponent.typeroof.jsx";
-import { _NOTDEF, _getFallback } from "./typeSpecGetDefaults.js";
+import { _NOTDEF, _getFallback } from "./typeSpecGetDefaults.mjs";
 import { StaticTag } from "../../generic.mjs";
 import {
     UITypeDrivenContainer,
     genericTypeToUIElement,
 } from "../../type-driven-ui.mjs";
-import { NODESPEC_PPS_MAP } from "./NODESPEC_PPS_MAP.js";
+import { NODESPEC_PPS_MAP } from "./NODESPEC_PPS_MAP.mjs";
 
 /**
  * Here's a good lesson, compared to typeSpecGetDefaults this is trivial,
