+++ lib/js/components/layouts/ramp/SyntheticValue.mjs
