+++ lib/js/components/layouts/ramp/TYPESPEC_PPS_MAP.mjs
