+++ lib/js/components/layouts/ramp/UIStylePatch.typeroof.jsx
@@ -4,13 +4,13 @@ import {
     UITypeDrivenContainer,
     genericTypeToUIElement,
 } from "../../type-driven-ui.mjs";
-import { typeSpecGetDefaults } from "./typeSpecGetDefaults.js";
+import { typeSpecGetDefaults } from "./typeSpecGetDefaults.mjs";
 import {
     SPECIFIC,
     ProcessedPropertiesSystemMap,
 } from "../../registered-properties-definitions.mjs";
 import { ForeignKey } from "../../../metamodel.mjs";
-import { TYPESPEC_PPS_MAP } from "./TYPESPEC_PPS_MAP.js";
+import { TYPESPEC_PPS_MAP } from "./TYPESPEC_PPS_MAP.mjs";
 import { getStylePatchFullLabel } from "../../type-spec-models.mjs";
 import { UICompositeStylePatch } from "./UICompositeStylePatch.typeroof.jsx";
