+++ lib/js/components/layouts/ramp/_getTypeSpecDefaultsMap.mjs
@@ -1,5 +1,5 @@
-import { SyntheticValue } from "./SyntheticValue.js";
-import { _pathSpecGen } from "./_pathSpecGen.js";
+import { SyntheticValue } from "./SyntheticValue.mjs";
+import { _pathSpecGen } from "./_pathSpecGen.mjs";
 import {
     ForeignKey,
     getFieldsByType,
@@ -25,12 +25,12 @@ import {
     LanguageTagModel,
     createLanguageTag,
 } from "../../language-tags.typeroof.jsx";
-import { fillTreeFromPaths } from "./fillTreeFromPaths.js";
+import { fillTreeFromPaths } from "./fillTreeFromPaths.mjs";
 import { identity, zip } from "../../../util.mjs";
 import { colorsGen, culoriToColor } from "../../color.mjs";
 import { runion_01_lineHeight } from "../../type-spec-fundamentals.mjs";
-import { TYPESPEC_PPS_MAP } from "./TYPESPEC_PPS_MAP.js";
-import { typeSpecGetDefaults } from "./typeSpecGetDefaults.js";
+import { TYPESPEC_PPS_MAP } from "./TYPESPEC_PPS_MAP.mjs";
+import { typeSpecGetDefaults } from "./typeSpecGetDefaults.mjs";
 import { LocalScopeTypeSpecnion } from "./_BaseTypeSpecnion.typeroof.jsx";
 
 // This will produce:
