+++ lib/js/components/layouts/ramp/_pathSpecGen.mjs
