+++ lib/js/components/layouts/ramp/typeSpecGetDefaults.mjs
